# src/main/resources/apis/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Event-Ease API",
    "description" : "Seamlessly manage any event quickly and easily with our versatile SaaS platform. Perfect for any occasion.",
    "termsOfService" : "Your terms here",
    "contact" : {
      "name" : "Example API Support",
      "url" : "http://exampleurl.com/contact",
      "email" : "coralo.luca@outlook.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/license/mit"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
  "paths" : {
...
    "/operation" : {
      "post" : {
        "tags" : [ "Operation Resource" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateOperationDTO"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OperationDTO"
                }
              }
            }
          }
        }
      }
    },
    "/operations" : {
      "get" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "activities",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }, {
          "name" : "endFrom",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }, {
          "name" : "events",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "operators",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }, {
          "name" : "startFrom",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }, {
          "name" : "statuses",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EventStatus"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OperationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/operations/{id}" : {
      "get" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OperationDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "head" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateOperationDTO"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OperationDTO"
                }
              }
            }
          }
        }
      }
    },
    "/operations/{id}/close" : {
      "post" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClosingReasonDTO"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        }
      }
    },
    "/operations/{id}/complete" : {
      "post" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        }
      }
    },
    "/operations/{id}/publish" : {
      "post" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        }
      }
    },
    "/operations:paged" : {
      "get" : {
        "tags" : [ "Operation Resource" ],
        "parameters" : [ {
          "name" : "activities",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }, {
          "name" : "endFrom",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }, {
          "name" : "events",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }, {
          "name" : "numPage",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "operators",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "startFrom",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }, {
          "name" : "statuses",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EventStatus"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginationOperationDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ClosingReasonDTO" : {
        "required" : [ "note" ],
        "type" : "object",
        "properties" : {
          "note" : {
            "maxLength" : 1000,
            "pattern" : "\\S",
            "type" : "string"
          }
        }
      },
...
      "CreateOperationDTO" : {
        "required" : [ "event", "activity", "startDateTime", "estimatedEndDateTime", "operator" ],
        "type" : "object",
        "properties" : {
          "event" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "activity" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "startDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "estimatedEndDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "startBookingDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "endBookingDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "location" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "operator" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "partecipants" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
...
      "EventStatus" : {
        "enum" : [ "DRAFT", "PUBLISHED", "IN_PROGRESS", "COMPLETED", "CANCELLED", "CLOSED" ],
        "type" : "string"
      },
      "Instant" : {
        "format" : "date-time",
        "type" : "string",
        "example" : "2022-03-10T16:15:50Z"
      },
      "OperationDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "event" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "activity" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "startBookingDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "endBookingDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "startDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "estimatedEndDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "endDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "status" : {
            "$ref" : "#/components/schemas/EventStatus"
          },
          "note" : {
            "type" : "string"
          },
          "location" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "operator" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "partecipants" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
...
      "PaginationOperationDTO" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OperationDTO"
            }
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalSize" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
      "UUID" : {
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
        "type" : "string"
      },
...
      "UpdateOperationDTO" : {
        "type" : "object",
        "properties" : {
          "event" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "activity" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "startDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "estimatedEndDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "startBookingDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "endBookingDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "location" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "operator" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "partecipants" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      }
    }
  }
}